Resolved duplicate classes, cleaned up code, added new doc comments.
